requirements.txt
# Core Dependencies
asyncio-mqtt>=0.16.2
sqlalchemy>=2.0.47
pyyaml>=6.0.3
tomlkit>=0.14.0
orjson>=3.11.7
loguru>=0.7.3
playwright>=1.58.0
rich>=14.3.3
rich-argparse>=1.7.2
psutil>=7.0.0
cryptography>=46.0.5

# Async Database Support
aiosqlite>=0.20.0

# Development Tools
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
ruff>=0.14.0
mypy>=1.10.0
